+++ plugins/markdown/plugin-markdown-preview/src/shared/locales.ts
@@ -1,0 +1,16 @@
+import type { ExactLocaleConfig } from '@vuepress/helper'
+
+export interface MarkdownPreviewPluginLocaleData {
+  /**
+   * Show code button text
+   */
+  show: string
+
+  /**
+   * Hide code button text
+   */
+  hide: string
+}
+
+export type MarkdownPreviewPluginLocaleConfig =
+  ExactLocaleConfig<MarkdownPreviewPluginLocaleData>